api/ai_engine/requirements.txt
moviepy
numpy
Pillow
opencv-python
pilmoji
emoji==1.7.0
requests
